refactor(cursor): remove transport toggles and simplify provider bootstrap
