--- cli/templates/devcontainer/sandcat/scripts/mitmproxy_addon.py
@@ -1,520 +1,0 @@
-"""
-mitmproxy addon: network access rules and secret substitution.
-
-Loaded via: mitmweb -s /scripts/mitmproxy_addon.py
-
-On startup, reads settings from up to three layers (lowest to highest
-precedence): user (~/.config/sandcat/settings.json), project
-(.sandcat/settings.json), and local (.sandcat/settings.local.json).
-Env vars and secrets are merged (higher precedence wins on conflict).
-Network rules are concatenated (highest precedence first).
-
-Network rules are evaluated top-to-bottom, first match wins, default deny.
-Secret placeholders are replaced with real values only for allowed hosts.
-"""
-
-import json
-import logging
-import os
-import re
-import subprocess
-import sys
-import base64
-import binascii
-from fnmatch import fnmatch
-
-from mitmproxy import ctx, http, dns
-
-_VALID_ENV_NAME = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*$")
-
-# Settings layers, lowest to highest precedence.
-SETTINGS_PATHS = [
-    "/config/settings.json",                # user:    ~/.config/sandcat/settings.json
-    "/config/project/settings.json",        # project: .sandcat/settings.json
-    "/config/project/settings.local.json",  # local:   .sandcat/settings.local.json
-]
-SANDCAT_ENV_PATH = "/home/mitmproxy/.mitmproxy/sandcat.env"
-SANDCAT_SECRETS_PATH = "/home/mitmproxy/.mitmproxy/sandcat-secrets.json"
-
-logger = logging.getLogger(__name__)
-
-class SandcatAddon:
-    def __init__(self):
-        self.secrets: dict[str, dict] = {}  # name -> {value, hosts, placeholder}
-        self.network_rules: list[dict] = []
-        self.env: dict[str, str] = {}  # non-secret env vars
-        self.debug_enabled = False
-        self.cursor_bearer_fallback_enabled = False
-
-    def load(self, loader):
-        layers = []
-        for path in SETTINGS_PATHS:
-            if os.path.isfile(path):
-                with open(path) as f:
-                    layers.append(json.load(f))
-
-        if not layers:
-            logger.info("No settings files found — addon disabled")
-            return
-
-        merged = self._merge_settings(layers)
-        raw_debug = merged.get("env", {}).get(
-            "SANDCAT_MITM_DEBUG", os.environ.get("SANDCAT_MITM_DEBUG", "")
-        )
-        self.debug_enabled = self._is_truthy(raw_debug)
-        raw_cursor_bearer_fallback = merged.get("env", {}).get(
-            "SANDCAT_CURSOR_BEARER_FALLBACK",
-            os.environ.get("SANDCAT_CURSOR_BEARER_FALLBACK", ""),
-        )
-        self.cursor_bearer_fallback_enabled = self._is_truthy(raw_cursor_bearer_fallback)
-
-        self._configure_op_token(merged.get("op_service_account_token"))
-        self.env = merged["env"]
-        self._load_secrets(merged["secrets"])
-        self._load_network_rules(merged["network"])
-        self._write_placeholders_env()
-        self._write_secrets_json()
-
-        ctx.log.info(
-            f"Loaded {len(self.env)} env var(s) and {len(self.secrets)} secret(s), "
-            f"wrote {SANDCAT_ENV_PATH} and {SANDCAT_SECRETS_PATH}"
-        )
-        self._debug(
-            "Debug mode enabled; set env.SANDCAT_MITM_DEBUG=true to keep this logging"
-        )
-
-    @staticmethod
-    def _is_truthy(value: str | bool | None) -> bool:
-        if value is None:
-            return False
-        if isinstance(value, bool):
-            return value
-        return str(value).strip().lower() in {"1", "true", "yes", "on"}
-
-    def _debug(self, message: str):
-        if self.debug_enabled:
-            ctx.log.info(f"[sandcat-debug] {message}")
-
-    @staticmethod
-    def _configure_op_token(token: str | None):
-        """Set OP_SERVICE_ACCOUNT_TOKEN from settings if not already in the environment."""
-        if token and "OP_SERVICE_ACCOUNT_TOKEN" not in os.environ:
-            os.environ["OP_SERVICE_ACCOUNT_TOKEN"] = token
-
-    @staticmethod
-    def _merge_settings(layers: list[dict]) -> dict:
-        """Merge settings from multiple layers (lowest to highest precedence).
-
-        - env: dict merge, higher precedence overwrites.
-        - secrets: dict merge, higher precedence overwrites.
-        - network: concatenated, highest precedence first.
-        - op_service_account_token: highest precedence non-empty value wins.
-        """
-        env: dict[str, str] = {}
-        secrets: dict[str, dict] = {}
-        network: list[dict] = []
-        op_token: str | None = None
-
-        for layer in layers:
-            env.update(layer.get("env", {}))
-            secrets.update(layer.get("secrets", {}))
-            layer_token = layer.get("op_service_account_token")
-            if layer_token:
-                op_token = layer_token
-
-        # Network rules: highest-precedence layer's rules come first.
-        for layer in reversed(layers):
-            network.extend(layer.get("network", []))
-
-        return {"env": env, "secrets": secrets, "network": network,
-                "op_service_account_token": op_token}
-
-    def _load_secrets(self, raw_secrets: dict):
-        for name, entry in raw_secrets.items():
-            placeholder = f"SANDCAT_PLACEHOLDER_{name}"
-            source = "op" if "op" in entry else "value"
-            try:
-                value = self._resolve_secret_value(name, entry)
-            except (RuntimeError, ValueError) as e:
-                ctx.log.warn(str(e))
-                print(f"WARNING: {e}", file=sys.stderr)
-                value = ""
-            self.secrets[name] = {
-                "value": value,
-                "hosts": entry.get("hosts", []),
-                "placeholder": placeholder,
-            }
-            self._debug(
-                f"Secret loaded: name={name} source={source} "
-                f"value_len={len(value)} hosts={entry.get('hosts', [])}"
-            )
-
-    @staticmethod
-    def _resolve_secret_value(name: str, entry: dict) -> str:
-        """Resolve a secret value from either a plain 'value' or a 1Password 'op' reference."""
-        has_value = "value" in entry
-        has_op = "op" in entry
-
-        if has_value and has_op:
-            raise ValueError(
-                f"Secret {name!r}: specify either 'value' or 'op', not both"
-            )
-        if not has_value and not has_op:
-            raise ValueError(
-                f"Secret {name!r}: must specify either 'value' or 'op'"
-            )
-
-        if has_value:
-            return entry["value"]
-
-        op_ref = entry["op"]
-        if not op_ref.startswith("op://"):
-            raise ValueError(
-                f"Secret {name!r}: 'op' value must start with 'op://', got {op_ref!r}"
-            )
-
-        try:
-            result = subprocess.run(
-                ["op", "read", op_ref],
-                capture_output=True, text=True, timeout=30,
-            )
-        except FileNotFoundError:
-            raise RuntimeError(
-                f"Secret {name!r}: 'op' CLI not found. "
-                "Install 1Password CLI to use op:// references."
-            ) from None
-
-        if result.returncode != 0:
-            stderr = result.stderr.strip()
-            raise RuntimeError(
-                f"Secret {name!r}: 'op read' failed: {stderr}"
-            )
-
-        return result.stdout.strip()
-
-    def _load_network_rules(self, raw_rules: list):
-        self.network_rules = raw_rules
-        ctx.log.info(f"Loaded {len(self.network_rules)} network rule(s)")
-
-    def _find_matching_rule(self, method: str | None, host: str) -> dict | None:
-        host = host.lower().rstrip(".")
-        for rule in self.network_rules:
-            if not fnmatch(host, rule["host"].lower()):
-                continue
-            rule_method = rule.get("method")
-            if (
-                rule_method is not None
-                and method is not None
-                and rule_method.upper() != method.upper()
-            ):
-                continue
-            return rule
-        return None
-
-    @staticmethod
-    def _shell_escape(value: str) -> str:
-        """Escape a string for safe inclusion inside double quotes in shell."""
-        return (value
-                .replace("\\", "\\\\")
-                .replace('"', '\\"')
-                .replace("$", "\\$")
-                .replace("`", "\\`")
-                .replace("\n", "\\n"))
-
-    @staticmethod
-    def _validate_env_name(name: str):
-        """Raise ValueError if name is not a valid shell variable name."""
-        if not _VALID_ENV_NAME.match(name):
-            raise ValueError(f"Invalid env var name: {name!r}")
-
-    def _write_placeholders_env(self):
-        lines = []
-        # Non-secret env vars (e.g. git identity) — passed through as-is.
-        for name, value in self.env.items():
-            self._validate_env_name(name)
-            lines.append(f'export {name}="{self._shell_escape(value)}"')
-        for name, entry in self.secrets.items():
-            self._validate_env_name(name)
-            lines.append(f'export {name}="{self._shell_escape(entry["placeholder"])}"')
-        with open(SANDCAT_ENV_PATH, "w") as f:
-            f.write("\n".join(lines) + "\n")
-
-    def _write_secrets_json(self):
-        """Write real secret values to a JSON file on the shared volume.
-
-        Used by agent init scripts for tools whose TLS implementation
-        prevents mitmproxy MITM (e.g. Cursor CLI cert pinning).
-        The file is mounted read-only in the agent container.
-        """
-        data = {name: entry["value"] for name, entry in self.secrets.items()}
-        with open(SANDCAT_SECRETS_PATH, "w") as f:
-            json.dump(data, f)
-        os.chmod(SANDCAT_SECRETS_PATH, 0o600)
-
-    def _is_request_allowed(self, method: str | None, host: str) -> bool:
-        rule = self._find_matching_rule(method, host)
-        if rule is None:
-            return False
-        return rule["action"] == "allow"
-
-    @staticmethod
-    def _is_cursor_host(host: str) -> bool:
-        host = host.lower().rstrip(".")
-        return (
-            host == "cursor.sh"
-            or host == "cursor.com"
-            or host.endswith(".cursor.sh")
-            or host.endswith(".cursor.com")
-        )
-
-    def _is_cursor_streaming_request(self, flow: http.HTTPFlow) -> bool:
-        host = flow.request.pretty_host.lower()
-        if not self._is_cursor_host(host):
-            return False
-
-        path = flow.request.path
-        if (
-            path.startswith("/agent.v1.AgentService/Run")
-            or path.startswith("/agent.v1.AgentService/RunSSE")
-            or path.startswith("/aiserver.v1.RepositoryService/")
-        ):
-            return True
-
-        content_type = flow.request.headers.get("content-type", "")
-        return "application/connect+proto" in content_type.lower()
-
-    @staticmethod
-    def _prepare_cursor_streaming_request(flow: http.HTTPFlow):
-        # Keep request transport opaque and uncompressed for connect/protobuf streams.
-        flow.request.headers["accept-encoding"] = "identity"
-        flow.request.stream = True
-
-    def _inject_cursor_streaming_auth(self, flow: http.HTTPFlow):
-        entry = self.secrets.get("CURSOR_API_KEY")
-        if not entry:
-            return
-
-        host = flow.request.pretty_host.lower()
-        allowed_hosts = entry.get("hosts", [])
-        if not any(fnmatch(host, pattern.lower()) for pattern in allowed_hosts):
-            flow.response = http.Response.make(
-                403,
-                f"Blocked: secret 'CURSOR_API_KEY' not allowed for host {host!r}\n".encode(),
-                {"Content-Type": "text/plain"},
-            )
-            ctx.log.warn(f"Blocked secret 'CURSOR_API_KEY' leak to disallowed host {host!r}")
-            return
-
-        value = entry.get("value", "")
-        if not value:
-            self._debug("Cursor streaming auth injection skipped: CURSOR_API_KEY is empty")
-            return
-
-        # Primary strategy selected by user: custom header injection.
-        flow.request.headers["CURSOR-API-KEY"] = value
-
-        # Optional fallback: add Bearer auth for endpoints that ignore custom header.
-        if self.cursor_bearer_fallback_enabled:
-            flow.request.headers["authorization"] = f"Bearer {value}"
-            self._debug("Cursor streaming auth injection: CURSOR-API-KEY + Bearer fallback")
-        else:
-            self._debug("Cursor streaming auth injection: CURSOR-API-KEY only")
-
-    @staticmethod
-    def _is_textual_content_type(content_type: str | None) -> bool:
-        if not content_type:
-            return False
-        media_type = content_type.split(";", 1)[0].strip().lower()
-        return (
-            media_type.startswith("text/")
-            or media_type in {
-                "application/json",
-                "application/x-www-form-urlencoded",
-                "application/xml",
-                "application/javascript",
-                "application/graphql",
-            }
-            or media_type.endswith("+json")
-            or media_type.endswith("+xml")
-        )
-
-    @staticmethod
-    def _basic_auth_contains_placeholder(auth_header: str | None, placeholder: str) -> bool:
-        if not auth_header:
-            return False
-        if not auth_header.lower().startswith("basic "):
-            return False
-        encoded = auth_header.split(" ", 1)[1].strip()
-        if not encoded:
-            return False
-        try:
-            decoded = base64.b64decode(encoded).decode("utf-8")
-        except (binascii.Error, UnicodeDecodeError, ValueError):
-            return False
-        return placeholder in decoded
-
-    @staticmethod
-    def _replace_placeholder_in_basic_auth(
-        auth_header: str | None, placeholder: str, value: str
-    ) -> tuple[str | None, bool]:
-        if not auth_header:
-            return auth_header, False
-        if not auth_header.lower().startswith("basic "):
-            return auth_header, False
-        encoded = auth_header.split(" ", 1)[1].strip()
-        if not encoded:
-            return auth_header, False
-        try:
-            decoded = base64.b64decode(encoded).decode("utf-8")
-        except (binascii.Error, UnicodeDecodeError, ValueError):
-            return auth_header, False
-        if placeholder not in decoded:
-            return auth_header, False
-        replaced = decoded.replace(placeholder, value)
-        new_encoded = base64.b64encode(replaced.encode("utf-8")).decode("ascii")
-        return f"Basic {new_encoded}", True
-
-    def _substitute_secrets(self, flow: http.HTTPFlow):
-        host = flow.request.pretty_host.lower()
-        is_cursor_streaming = self._is_cursor_streaming_request(flow)
-
-        for name, entry in self.secrets.items():
-            placeholder = entry["placeholder"]
-            value = entry["value"]
-            allowed_hosts = entry["hosts"]
-            auth_header = flow.request.headers.get("authorization", "")
-
-            present = (
-                placeholder in flow.request.url
-                or placeholder in str(flow.request.headers)
-                or self._basic_auth_contains_placeholder(auth_header, placeholder)
-                or (
-                    not is_cursor_streaming
-                    and flow.request.content
-                    and placeholder.encode() in flow.request.content
-                )
-            )
-
-            if not present:
-                continue
-
-            # Leak detection: block if secret going to disallowed host
-            if not any(fnmatch(host, pattern.lower()) for pattern in allowed_hosts):
-                flow.response = http.Response.make(
-                    403,
-                    f"Blocked: secret {name!r} not allowed for host {host!r}\n".encode(),
-                    {"Content-Type": "text/plain"},
-                )
-                ctx.log.warn(
-                    f"Blocked secret {name!r} leak to disallowed host {host!r}"
-                )
-                self._debug(
-                    f"Secret leak blocked: name={name} host={host} "
-                    f"allowed_hosts={allowed_hosts}"
-                )
-                return
-
-            if not value:
-                self._debug(
-                    f"Placeholder present but secret value empty: name={name} host={host}"
-                )
-
-            url_replaced = False
-            if placeholder in flow.request.url:
-                flow.request.url = flow.request.url.replace(placeholder, value)
-                url_replaced = True
-            header_replacements = 0
-            for k, v in flow.request.headers.items():
-                if placeholder in v:
-                    flow.request.headers[k] = v.replace(placeholder, value)
-                    header_replacements += 1
-            updated_auth, replaced_basic_auth = self._replace_placeholder_in_basic_auth(
-                flow.request.headers.get("authorization", ""), placeholder, value
-            )
-            if replaced_basic_auth and updated_auth is not None:
-                flow.request.headers["authorization"] = updated_auth
-                header_replacements += 1
-            body_replaced = False
-            body_skipped_binary = False
-            body_skipped_cursor = False
-            if is_cursor_streaming:
-                body_skipped_cursor = True
-            elif flow.request.content and placeholder.encode() in flow.request.content:
-                content_type = flow.request.headers.get("content-type", "")
-                if self._is_textual_content_type(content_type):
-                    flow.request.content = flow.request.content.replace(
-                        placeholder.encode(), value.encode()
-                    )
-                    body_replaced = True
-                else:
-                    # Do not mutate binary/protobuf payloads; length-delimited
-                    # framing can break and cause server-side parse failures.
-                    body_skipped_binary = True
-            self._debug(
-                f"Secret substitution: name={name} host={host} "
-                f"url={url_replaced} headers={header_replacements} "
-                f"body={body_replaced} body_skipped_binary={body_skipped_binary} "
-                f"body_skipped_cursor={body_skipped_cursor}"
-            )
-
-    def request(self, flow: http.HTTPFlow):
-        method = flow.request.method
-        host = flow.request.pretty_host
-
-        matching_rule = self._find_matching_rule(method, host)
-        if matching_rule is None:
-            self._debug(
-                f"Policy no-match -> deny: method={method} host={host}"
-            )
-        else:
-            self._debug(
-                f"Policy match: method={method} host={host} "
-                f"rule_host={matching_rule.get('host')} "
-                f"rule_method={matching_rule.get('method')} "
-                f"action={matching_rule.get('action')}"
-            )
-
-        if matching_rule is None or matching_rule.get("action") != "allow":
-            flow.response = http.Response.make(
-                403,
-                f"Blocked by network policy: {method} {host}\n".encode(),
-                {"Content-Type": "text/plain"},
-            )
-            ctx.log.warn(f"Network deny: {method} {host}")
-            return
-
-        if self._is_cursor_streaming_request(flow):
-            self._prepare_cursor_streaming_request(flow)
-            self._inject_cursor_streaming_auth(flow)
-            if flow.response is not None:
-                # _inject_cursor_streaming_auth may block on host leak check.
-                return
-
-        self._substitute_secrets(flow)
-
-    def responseheaders(self, flow: http.HTTPFlow):
-        if self._is_cursor_streaming_request(flow):
-            flow.response.stream = True
-
-    def dns_request(self, flow: dns.DNSFlow):
-        question = flow.request.question
-        if question is None:
-            flow.response = flow.request.fail(dns.response_codes.REFUSED)
-            return
-
-        host = question.name
-        matching_rule = self._find_matching_rule(None, host)
-        if matching_rule is None:
-            self._debug(f"DNS no-match -> deny: host={host}")
-        else:
-            self._debug(
-                f"DNS policy match: host={host} "
-                f"rule_host={matching_rule.get('host')} action={matching_rule.get('action')}"
-            )
-
-        if matching_rule is None or matching_rule.get("action") != "allow":
-            flow.response = flow.request.fail(dns.response_codes.REFUSED)
-            ctx.log.warn(f"DNS deny: {host}")
-
-
-addons = [SandcatAddon()]
