+++ cli/templates/devcontainer/sandcat/scripts/mitmproxy_addon_claude.py
@@ -1,9 +1,18 @@
 """
-Claude-focused mitmproxy addon: original network policy + secret substitution.
+mitmproxy addon: network access rules and secret substitution.
+
+Loaded via: mitmweb -s /scripts/mitmproxy_addon.py
+
+On startup, reads settings from up to three layers (lowest to highest
+precedence): user (~/.config/sandcat/settings.json), project
+(.sandcat/settings.json), and local (.sandcat/settings.local.json).
+Env vars and secrets are merged (higher precedence wins on conflict).
+Network rules are concatenated (highest precedence first).
+
+Network rules are evaluated top-to-bottom, first match wins, default deny.
+Secret placeholders are replaced with real values only for allowed hosts.
 """
 
-import base64
-import binascii
 import json
 import logging
 import os
@@ -12,27 +21,25 @@ import subprocess
 import sys
 from fnmatch import fnmatch
 
-from mitmproxy import ctx, dns, http
+from mitmproxy import ctx, http, dns
 
 _VALID_ENV_NAME = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*$")
 
+# Settings layers, lowest to highest precedence.
 SETTINGS_PATHS = [
-    "/config/settings.json",
-    "/config/project/settings.json",
-    "/config/project/settings.local.json",
+    "/config/settings.json",                # user:    ~/.config/sandcat/settings.json
+    "/config/project/settings.json",        # project: .sandcat/settings.json
+    "/config/project/settings.local.json",  # local:   .sandcat/settings.local.json
 ]
 SANDCAT_ENV_PATH = "/home/mitmproxy/.mitmproxy/sandcat.env"
-SANDCAT_SECRETS_PATH = "/home/mitmproxy/.mitmproxy/sandcat-secrets.json"
 
 logger = logging.getLogger(__name__)
 
-
 class SandcatAddon:
     def __init__(self):
-        self.secrets: dict[str, dict] = {}
+        self.secrets: dict[str, dict] = {}  # name -> {value, hosts, placeholder}
         self.network_rules: list[dict] = []
-        self.env: dict[str, str] = {}
-        self.debug_enabled = False
+        self.env: dict[str, str] = {}  # non-secret env vars
 
     def load(self, loader):
         layers = []
@@ -46,38 +53,32 @@ class SandcatAddon:
             return
 
         merged = self._merge_settings(layers)
-        raw_debug = merged.get("env", {}).get(
-            "SANDCAT_MITM_DEBUG", os.environ.get("SANDCAT_MITM_DEBUG", "")
-        )
-        self.debug_enabled = self._is_truthy(raw_debug)
 
         self._configure_op_token(merged.get("op_service_account_token"))
         self.env = merged["env"]
         self._load_secrets(merged["secrets"])
         self._load_network_rules(merged["network"])
         self._write_placeholders_env()
-        self._write_secrets_json()
 
         ctx.log.info(
-            f"Loaded {len(self.env)} env var(s) and {len(self.secrets)} secret(s), "
-            f"wrote {SANDCAT_ENV_PATH} and {SANDCAT_SECRETS_PATH}"
+            f"Loaded {len(self.env)} env var(s) and {len(self.secrets)} secret(s), wrote {SANDCAT_ENV_PATH}"
         )
 
     @staticmethod
-    def _is_truthy(value: str | bool | None) -> bool:
-        if value is None:
-            return False
-        if isinstance(value, bool):
-            return value
-        return str(value).strip().lower() in {"1", "true", "yes", "on"}
-
-    @staticmethod
     def _configure_op_token(token: str | None):
+        """Set OP_SERVICE_ACCOUNT_TOKEN from settings if not already in the environment."""
         if token and "OP_SERVICE_ACCOUNT_TOKEN" not in os.environ:
             os.environ["OP_SERVICE_ACCOUNT_TOKEN"] = token
 
     @staticmethod
     def _merge_settings(layers: list[dict]) -> dict:
+        """Merge settings from multiple layers (lowest to highest precedence).
+
+        - env: dict merge, higher precedence overwrites.
+        - secrets: dict merge, higher precedence overwrites.
+        - network: concatenated, highest precedence first.
+        - op_service_account_token: highest precedence non-empty value wins.
+        """
         env: dict[str, str] = {}
         secrets: dict[str, dict] = {}
         network: list[dict] = []
@@ -90,15 +91,12 @@ class SandcatAddon:
             if layer_token:
                 op_token = layer_token
 
+        # Network rules: highest-precedence layer's rules come first.
         for layer in reversed(layers):
             network.extend(layer.get("network", []))
 
-        return {
-            "env": env,
-            "secrets": secrets,
-            "network": network,
-            "op_service_account_token": op_token,
-        }
+        return {"env": env, "secrets": secrets, "network": network,
+            "op_service_account_token": op_token}
 
     def _load_secrets(self, raw_secrets: dict):
         for name, entry in raw_secrets.items():
@@ -117,62 +115,70 @@ class SandcatAddon:
 
     @staticmethod
     def _resolve_secret_value(name: str, entry: dict) -> str:
+        """Resolve a secret value from either a plain 'value' or a 1Password 'op' reference."""
         has_value = "value" in entry
         has_op = "op" in entry
+
         if has_value and has_op:
-            raise ValueError(f"Secret {name!r}: specify either 'value' or 'op', not both")
+            raise ValueError(
+                f"Secret {name!r}: specify either 'value' or 'op', not both"
+            )
         if not has_value and not has_op:
-            raise ValueError(f"Secret {name!r}: must specify either 'value' or 'op'")
+            raise ValueError(
+                f"Secret {name!r}: must specify either 'value' or 'op'"
+            )
+
         if has_value:
             return entry["value"]
 
         op_ref = entry["op"]
         if not op_ref.startswith("op://"):
-            raise ValueError(f"Secret {name!r}: 'op' value must start with 'op://', got {op_ref!r}")
+            raise ValueError(
+                f"Secret {name!r}: 'op' value must start with 'op://', got {op_ref!r}"
+            )
 
         try:
-            result = subprocess.run(["op", "read", op_ref], capture_output=True, text=True, timeout=30)
+            result = subprocess.run(
+                ["op", "read", op_ref],
+                capture_output=True, text=True, timeout=30,
+                )
         except FileNotFoundError:
             raise RuntimeError(
-                f"Secret {name!r}: 'op' CLI not found. Install 1Password CLI to use op:// references."
+                f"Secret {name!r}: 'op' CLI not found. "
+                "Install 1Password CLI to use op:// references."
             ) from None
 
         if result.returncode != 0:
-            raise RuntimeError(f"Secret {name!r}: 'op read' failed: {result.stderr.strip()}")
+            stderr = result.stderr.strip()
+            raise RuntimeError(
+                f"Secret {name!r}: 'op read' failed: {stderr}"
+            )
+
         return result.stdout.strip()
 
     def _load_network_rules(self, raw_rules: list):
         self.network_rules = raw_rules
         ctx.log.info(f"Loaded {len(self.network_rules)} network rule(s)")
 
-    def _find_matching_rule(self, method: str | None, host: str) -> dict | None:
-        host = host.lower().rstrip(".")
-        for rule in self.network_rules:
-            if not fnmatch(host, rule["host"].lower()):
-                continue
-            rule_method = rule.get("method")
-            if rule_method is not None and method is not None and rule_method.upper() != method.upper():
-                continue
-            return rule
-        return None
-
     @staticmethod
     def _shell_escape(value: str) -> str:
-        return (
-            value.replace("\\", "\\\\")
+        """Escape a string for safe inclusion inside double quotes in shell."""
+        return (value
+            .replace("\\", "\\\\")
             .replace('"', '\\"')
             .replace("$", "\\$")
             .replace("`", "\\`")
-            .replace("\n", "\\n")
-        )
+            .replace("\n", "\\n"))
 
     @staticmethod
     def _validate_env_name(name: str):
+        """Raise ValueError if name is not a valid shell variable name."""
         if not _VALID_ENV_NAME.match(name):
             raise ValueError(f"Invalid env var name: {name!r}")
 
     def _write_placeholders_env(self):
         lines = []
+        # Non-secret env vars (e.g. git identity) — passed through as-is.
         for name, value in self.env.items():
             self._validate_env_name(name)
             lines.append(f'export {name}="{self._shell_escape(value)}"')
@@ -182,116 +188,64 @@ class SandcatAddon:
         with open(SANDCAT_ENV_PATH, "w") as f:
             f.write("\n".join(lines) + "\n")
 
-    def _write_secrets_json(self):
-        data = {name: entry["value"] for name, entry in self.secrets.items()}
-        with open(SANDCAT_SECRETS_PATH, "w") as f:
-            json.dump(data, f)
-        os.chmod(SANDCAT_SECRETS_PATH, 0o600)
-
-    @staticmethod
-    def _is_textual_content_type(content_type: str | None) -> bool:
-        if not content_type:
-            return False
-        media_type = content_type.split(";", 1)[0].strip().lower()
-        return (
-            media_type.startswith("text/")
-            or media_type
-            in {
-                "application/json",
-                "application/x-www-form-urlencoded",
-                "application/xml",
-                "application/javascript",
-                "application/graphql",
-            }
-            or media_type.endswith("+json")
-            or media_type.endswith("+xml")
-        )
-
-    @staticmethod
-    def _basic_auth_contains_placeholder(auth_header: str | None, placeholder: str) -> bool:
-        if not auth_header or not auth_header.lower().startswith("basic "):
-            return False
-        encoded = auth_header.split(" ", 1)[1].strip()
-        if not encoded:
-            return False
-        try:
-            decoded = base64.b64decode(encoded).decode("utf-8")
-        except (binascii.Error, UnicodeDecodeError, ValueError):
-            return False
-        return placeholder in decoded
-
-    @staticmethod
-    def _replace_placeholder_in_basic_auth(
-        auth_header: str | None, placeholder: str, value: str
-    ) -> tuple[str | None, bool]:
-        if not auth_header or not auth_header.lower().startswith("basic "):
-            return auth_header, False
-        encoded = auth_header.split(" ", 1)[1].strip()
-        if not encoded:
-            return auth_header, False
-        try:
-            decoded = base64.b64decode(encoded).decode("utf-8")
-        except (binascii.Error, UnicodeDecodeError, ValueError):
-            return auth_header, False
-        if placeholder not in decoded:
-            return auth_header, False
-        replaced = decoded.replace(placeholder, value)
-        return f"Basic {base64.b64encode(replaced.encode('utf-8')).decode('ascii')}", True
+    def _is_request_allowed(self, method: str | None, host: str) -> bool:
+        host = host.lower().rstrip(".")
+        for rule in self.network_rules:
+            if not fnmatch(host, rule["host"].lower()):
+                continue
+            rule_method = rule.get("method")
+            if rule_method is not None and method is not None and rule_method.upper() != method.upper():
+                continue
+            return rule["action"] == "allow"
+        return False # default deny
 
     def _substitute_secrets(self, flow: http.HTTPFlow):
         host = flow.request.pretty_host.lower()
+
         for name, entry in self.secrets.items():
             placeholder = entry["placeholder"]
             value = entry["value"]
             allowed_hosts = entry["hosts"]
-            auth_header = flow.request.headers.get("authorization", "")
 
             present = (
                 placeholder in flow.request.url
                 or placeholder in str(flow.request.headers)
-                or self._basic_auth_contains_placeholder(auth_header, placeholder)
                 or (
                     flow.request.content
                     and placeholder.encode() in flow.request.content
                 )
             )
+ 
             if not present:
                 continue
 
+            # Leak detection: block if secret going to disallowed host
             if not any(fnmatch(host, pattern.lower()) for pattern in allowed_hosts):
                 flow.response = http.Response.make(
                     403,
                     f"Blocked: secret {name!r} not allowed for host {host!r}\n".encode(),
                     {"Content-Type": "text/plain"},
                 )
-                ctx.log.warn(f"Blocked secret {name!r} leak to disallowed host {host!r}")
+                ctx.log.warn(
+                    f"Blocked secret {name!r} leak to disallowed host {host!r}"
+                )
                 return
 
             if placeholder in flow.request.url:
                 flow.request.url = flow.request.url.replace(placeholder, value)
-
             for k, v in flow.request.headers.items():
                 if placeholder in v:
                     flow.request.headers[k] = v.replace(placeholder, value)
-
-            updated_auth, replaced_basic_auth = self._replace_placeholder_in_basic_auth(
-                flow.request.headers.get("authorization", ""), placeholder, value
-            )
-            if replaced_basic_auth and updated_auth is not None:
-                flow.request.headers["authorization"] = updated_auth
-
             if flow.request.content and placeholder.encode() in flow.request.content:
-                content_type = flow.request.headers.get("content-type", "")
-                if self._is_textual_content_type(content_type):
-                    flow.request.content = flow.request.content.replace(
-                        placeholder.encode(), value.encode()
-                    )
+                flow.request.content = flow.request.content.replace(
+                    placeholder.encode(), value.encode()
+                )
 
     def request(self, flow: http.HTTPFlow):
         method = flow.request.method
         host = flow.request.pretty_host
-        matching_rule = self._find_matching_rule(method, host)
-        if matching_rule is None or matching_rule.get("action") != "allow":
+
+        if not self._is_request_allowed(method, host):
             flow.response = http.Response.make(
                 403,
                 f"Blocked by network policy: {method} {host}\n".encode(),
@@ -299,6 +253,7 @@ class SandcatAddon:
             )
             ctx.log.warn(f"Network deny: {method} {host}")
             return
+        
         self._substitute_secrets(flow)
 
     def dns_request(self, flow: dns.DNSFlow):
@@ -306,12 +261,11 @@ class SandcatAddon:
         if question is None:
             flow.response = flow.request.fail(dns.response_codes.REFUSED)
             return
+
         host = question.name
-        matching_rule = self._find_matching_rule(None, host)
-        if matching_rule is None or matching_rule.get("action") != "allow":
+        if not self._is_request_allowed(None, host):
             flow.response = flow.request.fail(dns.response_codes.REFUSED)
             ctx.log.warn(f"DNS deny: {host}")
 
 
 addons = [SandcatAddon()]
-
